feat(ranking): 월간 및 주간 랭킹 엔티티 수정
- rankPosition 필드 타입을 int에서 long으로 변경
- createdAt 필드 타입을 ZonedDateTime에서 LocalDateTime으로 변경
- 불필요한 메서드 주석 제거

+++ modules/jpa/src/main/java/com/loopers/domain/ranking/MonthlyRankEntity.java
@@ -46,7 +46,7 @@ public class MonthlyRankEntity {
     private long totalScore;
 
     @Column(name = "rank_position", nullable = false)
-    private int rankPosition;
+    private long rankPosition;
 
     @Column(name = "created_at", nullable = false)
     private LocalDateTime createdAt;
